requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
node==1.2.2
odict==1.9.0
plumber==1.7
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.4.post2
zope.lifecycleevent==5.0
zope.proxy==5.2
